// src/interfaces/Responses.ts
export interface IResponses {
  onAskResponses: string[];
  onFetchQuoteResponses: string[];
  onBalanceResponses: string[];
  onDailiesResponses: string[];
  onSpinResponses: string[];
  onAssignRoleResponses: string[];
  onRemoveRoleResponses: string[];
  onFetchChapterResponses: string[];
  onHelpResponses: string[];
}